requirements.txt
django
python-dotenv
requests
redis
celery
django-celery-beat